+++ tests/utils.py
@@ -871,10 +871,6 @@ class TestFiles:
     def test_unsupported_requirements_file(self):
         return os.path.join(self.resources_path, "test_requirements_unsupported.txt")
 
-    @property
-    def test_conda_environment_file(self):
-        return os.path.join(self.resources_path, "test_environment.yml")
-
 
 class TypeMap(NamedTuple):
     col_name: str