+++ CHANGELOG.md
@@ -2,13 +2,24 @@
 
 ## 1.6.0 (TBD)
 
-### New Features
+## New Features
 
+- Added support for new functions in `snowflake.snowpark.functions`:
+  - `array_sort`
+  - `sort_array`
+  - `array_min`
+  - `array_max`
 - Added support for Python packages specified via `Session.add_requirements` or `Session.add_packages`, now usable even if packages are not present on the Snowflake Anaconda channel.
+- Added support for async execution of multi-query dataframe containing binding variables.
+- Added support for renaming multiple columns in `DataFrame.rename`.
 
 ### Improvements
 
 - Removed redundant dependency `typing-extensions`.
+
+### Dependency updates
+
+- Updated ``snowflake-connector-python`` to 3.0.4.
 
 ## 1.5.1 (2023-06-20)